Rename IsLoggedIn to LoginStatus https://bugs.webkit.org/show_bug.cgi?id=275974
Reviewed by Charlie Wolfe.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/DerivedSources-input.xcfilelist:
* Source/WebCore/DerivedSources-output.xcfilelist:
* Source/WebCore/DerivedSources.make:
* Source/WebCore/Headers.cmake:
* Source/WebCore/SmartPointerExpectations/NoUncountedMemberCheckerExpectations:
* Source/WebCore/SmartPointerExpectations/UncountedCallArgsCheckerExpectations:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/page/LoginStatus.cpp: Renamed from Source/WebCore/page/LoggedInStatus.cpp.
(WebCore::LoginStatus::create):
(WebCore::LoginStatus::LoginStatus):
(WebCore::LoginStatus::setTimeToLive):
(WebCore::LoginStatus::hasExpired const):
(WebCore::LoginStatus::expiry const):
* Source/WebCore/page/LoginStatus.h: Renamed from Source/WebCore/page/LoggedInStatus.h.
(WebCore::LoginStatus::registrableDomain const):
(WebCore::LoginStatus::username const):
(WebCore::LoginStatus::credentialTokenType const):
(WebCore::LoginStatus::authenticationType const):
(WebCore::LoginStatus::loggedInTime const):
* Source/WebCore/page/Navigator+LoginStatus.idl: Renamed from Source/WebCore/page/Navigator+IsLoggedIn.idl.
* Source/WebCore/page/NavigatorLoginStatus.cpp: Renamed from Source/WebCore/page/NavigatorIsLoggedIn.cpp.
(WebCore::NavigatorLoginStatus::from):
(WebCore::NavigatorLoginStatus::supplementName):
(WebCore::NavigatorLoginStatus::setLoggedIn):
(WebCore::NavigatorLoginStatus::setLoggedOut):
(WebCore::NavigatorLoginStatus::isLoggedIn):
* Source/WebCore/page/NavigatorLoginStatus.h: Renamed from Source/WebCore/page/NavigatorIsLoggedIn.h.
* Tools/DumpRenderTree/TestOptions.cpp:
(WTR::TestOptions::defaults):
* Tools/TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* Tools/TestWebKitAPI/Tests/WebCore/LoggedInStatus.cpp: Removed.
* Tools/TestWebKitAPI/Tests/WebCore/LoginStatus.cpp: Added.
(TestWebKitAPI::TEST(LoginStatus, DefaultExpiryWebAuthn)):
(TestWebKitAPI::TEST(LoginStatus, DefaultExpiryPasswordManager)):
(TestWebKitAPI::TEST(LoginStatus, DefaultExpiryUnmanaged)):
(TestWebKitAPI::TEST(LoginStatus, CustomExpiryBelowLong)):
(TestWebKitAPI::TEST(LoginStatus, CustomExpiryBelowShort)):
(TestWebKitAPI::TEST(LoginStatus, RenewedExpiry)):
(TestWebKitAPI::TEST(LoginStatus, InvalidUsernames)):
(TestWebKitAPI::TEST(LoginStatus, ClampedExpiryLong)):
(TestWebKitAPI::TEST(LoginStatus, ClampedExpiryShort)):

Canonical link: https://commits.webkit.org/280497@main
