+++ Source/WebCore/DerivedSources-input.xcfilelist
@@ -1786,7 +1786,7 @@ $(PROJECT_DIR)/page/NavigationHistoryEntry.idl
 $(PROJECT_DIR)/page/NavigationInterceptHandler.idl
 $(PROJECT_DIR)/page/NavigationNavigationType.idl
 $(PROJECT_DIR)/page/NavigationTransition.idl
-$(PROJECT_DIR)/page/Navigator+IsLoggedIn.idl
+$(PROJECT_DIR)/page/Navigator+LoginStatus.idl
 $(PROJECT_DIR)/page/Navigator+UserActivation.idl
 $(PROJECT_DIR)/page/Navigator.idl
 $(PROJECT_DIR)/page/NavigatorCookies.idl
