+++ Source/WebCore/SmartPointerExpectations/UncountedCallArgsCheckerExpectations
@@ -1507,7 +1507,7 @@ page/NavigationDestination.h
 page/NavigationHistoryEntry.cpp
 page/Navigator.cpp
 page/NavigatorBase.cpp
-page/NavigatorIsLoggedIn.cpp
+page/NavigatorLoginStatus.cpp
 page/OpportunisticTaskScheduler.cpp
 page/Page.cpp
 page/Page.h
