+++ Source/WebCore/Sources.txt
@@ -1991,7 +1991,7 @@ page/LocalFrame.cpp
 page/LocalFrameView.cpp
 page/LocalFrameViewLayoutContext.cpp
 page/Location.cpp
-page/LoggedInStatus.cpp
+page/LoginStatus.cpp
 page/MemoryRelease.cpp
 page/MouseEventWithHitTestResults.cpp
 page/NavigateEvent.cpp
@@ -2003,7 +2003,7 @@ page/NavigationHistoryEntry.cpp
 page/NavigationTransition.cpp
 page/Navigator.cpp
 page/NavigatorBase.cpp
-page/NavigatorIsLoggedIn.cpp
+page/NavigatorLoginStatus.cpp
 page/OpportunisticTaskScheduler.cpp
 page/OriginAccessEntry.cpp
 page/OriginAccessPatterns.cpp
