+++ Source/WebCore/page/NavigatorLoginStatus.cpp
@@ -1,0 +1,86 @@
+/*
+ * Copyright (C) 2019 Apple Inc. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS''
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS
+ * BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
+ * THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "config.h"
+#include "NavigatorLoginStatus.h"
+
+#include "JSDOMPromiseDeferred.h"
+#include "Navigator.h"
+
+namespace WebCore {
+
+NavigatorLoginStatus* NavigatorLoginStatus::from(Navigator& navigator)
+{
+    auto* supplement = static_cast<NavigatorLoginStatus*>(Supplement<Navigator>::from(&navigator, supplementName()));
+    if (!supplement) {
+        auto newSupplement = makeUnique<NavigatorLoginStatus>(navigator);
+        supplement = newSupplement.get();
+        provideTo(&navigator, supplementName(), WTFMove(newSupplement));
+    }
+    return supplement;
+}
+
+ASCIILiteral NavigatorLoginStatus::supplementName()
+{
+    return "NavigatorLoginStatus"_s;
+}
+
+void NavigatorLoginStatus::setLoggedIn(Navigator& navigator, Ref<DeferredPromise>&& promise)
+{
+    NavigatorLoginStatus::from(navigator)->setLoggedIn(WTFMove(promise));
+}
+
+void NavigatorLoginStatus::setLoggedOut(Navigator& navigator, Ref<DeferredPromise>&& promise)
+{
+    NavigatorLoginStatus::from(navigator)->setLoggedOut(WTFMove(promise));
+}
+
+void NavigatorLoginStatus::isLoggedIn(Navigator& navigator, Ref<DeferredPromise>&& promise)
+{
+    NavigatorLoginStatus::from(navigator)->isLoggedIn(WTFMove(promise));
+}
+
+void NavigatorLoginStatus::setLoggedIn(Ref<DeferredPromise>&& promise)
+{
+    if (m_navigator.cookieEnabled())
+        promise->resolve();
+    else
+        promise->reject();
+}
+
+void NavigatorLoginStatus::setLoggedOut(Ref<DeferredPromise>&& promise)
+{
+    promise->resolve();
+}
+
+void NavigatorLoginStatus::isLoggedIn(Ref<DeferredPromise>&& promise)
+{
+    if (m_navigator.cookieEnabled())
+        promise->resolve<IDLBoolean>(true);
+    else
+        promise->resolve<IDLBoolean>(false);
+}
+
+} // namespace WebCore
